// src/modules/common/enums/day-of-week.enum.ts
export enum DAY_OF_WEEK_ENUM {
  SUNDAY = 0,
  MONDAY = 1,
  TUESDAY = 2,
  WEDNESDAY = 3,
  THURSDAY = 4,
  FRIDAY = 5,
  SATURDAY = 6,
}

export type DayOfWeekType =
  | DAY_OF_WEEK_ENUM.SUNDAY
  | DAY_OF_WEEK_ENUM.MONDAY
  | DAY_OF_WEEK_ENUM.TUESDAY
  | DAY_OF_WEEK_ENUM.WEDNESDAY
  | DAY_OF_WEEK_ENUM.THURSDAY
  | DAY_OF_WEEK_ENUM.FRIDAY
  | DAY_OF_WEEK_ENUM.SATURDAY;