// src/modules/common/enums/day-of-week-short-name.enum.ts
export enum DAY_OF_WEEK_SHORT_NAME_ENUM {
  SUNDAY = 'sun',
  MONDAY = 'mon',
  TUESDAY = 'tue',
  WEDNESDAY = 'wed',
  THURSDAY = 'thu',
  FRIDAY = 'fri',
  SATURDAY = 'sat',
}

export type DayOfWeekShortNameType =
  | DAY_OF_WEEK_SHORT_NAME_ENUM.SUNDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.MONDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.TUESDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.WEDNESDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.THURSDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.FRIDAY
  | DAY_OF_WEEK_SHORT_NAME_ENUM.SATURDAY;